ControlBlock comments in C

+++ software/header/sha256.h
@@ -9,6 +9,6 @@ const uint32_t K[], H[], chunk_bits_count, max_chunks_count, chunk_word_count, c
 uint32_t left_rotate(uint32_t n, uint32_t d);
 uint32_t right_rotate(uint32_t n, uint32_t d);
 
-extern char* compute_sha256(char* data, int len);
+extern char* compute_sha256(char* data, uint8_t len);
 
 #endif
